terminal_based_project/whatsapp_agent/pyproject.toml
[project]
name = "whatsapp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
]
